requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
django==4.1.5
django-extensions==3.2.1
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3 ; python_version >= '3.5'
